WIP Attempting to flatten the domain tree by pushing annotations down into the typed value domains.
This adds at least one more domain per data-type.
That could be avoided in a crisp manner if we had ? domains (absence meaning null).

It would also require data-type-specific annot/content accessors.
That wouldn't be bad if their creation was automatic/implicit in the DSL.

I do like the `Stx` prefix in place of `R`, I should push that down.