WIP Adding varargs `seq_concat`
Hard to use when there might be zero arguments, how to construct the result?
Would be nice to have an "empty sequence" value that can lift to all sequence domains.

This isn't working out as anticipated.
Was hoping it would combine well with `map` eg (apply concat (map ...)) as an alternative to `fold`.
But it currently requires a sentinel sequence to handle the empty case.

Perhaps we can accept a sequence-domain as first argument?  How would the printer handle that?